Add a `JacobiPreconditioner` class
The new class returns a callable object that can be used as
preconditioner: it generates a Jacobi preconditioner for a given
objective function. This class can be used to update the preconditioner
on every iteration through an inversion.
Create a new `preconditioner.py` file where the new class and the
`get_jacobi_preconditioner` function live.
