+++ src/inversion_ideas/__init__.py
@@ -12,6 +12,7 @@ from .errors import ConvergenceWarning
 from .inversion import Inversion
 from .inversion_log import InversionLog, InversionLogRich
 from .minimizers import ConjugateGradient
+from .preconditioners import JacobiPreconditioner, get_jacobi_preconditioner
 from .regularization import TikhonovZero
 
 __all__ = [
@@ -23,6 +24,7 @@ __all__ = [
     "Inversion",
     "InversionLog",
     "InversionLogRich",
+    "JacobiPreconditioner",
     "ModelChanged",
     "MultiplierCooler",
     "ObjectiveChanged",
@@ -30,5 +32,6 @@ __all__ = [
     "__version__",
     "base",
     "create_inversion",
+    "get_jacobi_preconditioner",
     "utils",
 ]
